requirements.txt
flask
pytesseract
Pillow
opencv-python
PyPDF2
pdf2image
